+++ src/racer/Vector.zig
@@ -21,31 +21,31 @@ pub const Vec3_Set: *fn (out: *Vec3, x: f32, y: f32, z: f32) callconv(.C) void =
     @ptrFromInt(0x42F7B0);
 pub const Vec3_Copy: *fn (out: *Vec3, in: *const Vec3) callconv(.C) void =
     @ptrFromInt(0x42F7D0);
-pub const Vec3_Eql: *fn (in1: *Vec3, in2: *Vec3) callconv(.C) bool =
+pub const Vec3_Eql: *fn (in1: *const Vec3, in2: *const Vec3) callconv(.C) bool =
     @ptrFromInt(0x42F7F0);
 pub const Vec3_Add: *fn (out: *Vec3, in1: *const Vec3, in2: *const Vec3) callconv(.C) void =
     @ptrFromInt(0x42F830);
 pub const Vec3_Sub: *fn (out: *Vec3, in1: *const Vec3, in2: *const Vec3) callconv(.C) void =
     @ptrFromInt(0x42F860);
-pub const Vec3_Dot: *fn (in1: *Vec3, in2: *Vec3) callconv(.C) f32 =
+pub const Vec3_Dot: *fn (in1: *const Vec3, in2: *const Vec3) callconv(.C) f32 =
     @ptrFromInt(0x42F890);
 pub const Vec3_Mag: *fn (in: *const Vec3) callconv(.C) f32 =
     @ptrFromInt(0x42F8C0);
-pub const Vec3_DistSq: *fn (in: *Vec3) callconv(.C) f32 =
+pub const Vec3_DistSq: *fn (in1: *const Vec3, in2: *const Vec3) callconv(.C) f32 =
     @ptrFromInt(0x42F910);
-pub const Vec3_Dist: *fn (in: *Vec3) callconv(.C) f32 =
+pub const Vec3_Dist: *fn (in1: *const Vec3, in2: *const Vec3) callconv(.C) f32 =
     @ptrFromInt(0x42F950);
-pub const Vec3_MulMat4x4: *fn (out: *Vec3, in1: *Vec3, in2: *Mat4x4) callconv(.C) void =
+pub const Vec3_MulMat4x4: *fn (out: *Vec3, in1: *const Vec3, in2: *const Mat4x4) callconv(.C) void =
     @ptrFromInt(0x430980);
 pub const Vec3_Norm: *fn (out: *Vec3) callconv(.C) void =
     @ptrFromInt(0x42F9B0);
-pub const Vec3_Cross: *fn (out: *Vec3, in1: *Vec3, in2: *Vec3) callconv(.C) void =
+pub const Vec3_Cross: *fn (out: *Vec3, in1: *const Vec3, in2: *const Vec3) callconv(.C) void =
     @ptrFromInt(0x42F9F0);
-pub const Vec3_Scale: *fn (out: *Vec3, sc: f32, in: *Vec3) callconv(.C) void =
+pub const Vec3_Scale: *fn (out: *Vec3, sc: f32, in: *const Vec3) callconv(.C) void =
     @ptrFromInt(0x42FA50);
-pub const Vec3_AddScale1: *fn (out: *Vec3, in1: *Vec3, sc2: f32, in2: *Vec3) callconv(.C) void =
+pub const Vec3_AddScale1: *fn (out: *Vec3, in1: *const Vec3, sc2: f32, in2: *const Vec3) callconv(.C) void =
     @ptrFromInt(0x42FA80);
-pub const Vec3_AddScale2: *fn (out: *Vec3, sc1: f32, in1: *Vec3, sc2: f32, in2: *Vec3) callconv(.C) void =
+pub const Vec3_AddScale2: *fn (out: *Vec3, sc1: f32, in1: *const Vec3, sc2: f32, in2: *const Vec3) callconv(.C) void =
     @ptrFromInt(0x42FAC0);
 
 // GAME CONSTANTS